Adding VR Editor
Fix invalid bindings for the editor action set

Configuring the VR Editor for standalone devices

Update the gradle build config to include the openxr editor apks

Configuring the VR Editor for standalone devices

Add a Meta OS splash screen for the Godot XR Editor

- Leverages the Meta OS splash screen functionality: https://developer.oculus.com/documentation/native/android/mobile-splash/
- Fix rebase merge issues

Add scrolling capability to the controller pointer

Fix issue causing tracking to break when running `DocTools::generate(...)`

The issue is caused by the creation of a second `OpenXRInterface` instance. This previously caused the `openxr_api` instance to override its `xr_interface` reference with an instance that was not completely initialized.
The fix consisted in moving the `openxr_api` instance setup within `OpenXRInterface::initialize()` ensuring that the `openxr_api` instance is configured with an initialized instance of `OpenXRInterface`.