Fix issue causing tracking to break when running `DocTools::generate(...)`
The issue is caused by the creation of a second `OpenXRInterface` instance. This previously caused the `openxr_api` instance to override its `xr_interface` reference with an instance that was not completely initialized.
The fix consisted in moving the `openxr_api` instance setup within `OpenXRInterface::initialize()` ensuring that the `openxr_api` instance is configured with an initialized instance of `OpenXRInterface`.